[TEST] - Move profile schemas to separate folder
