+++ package.json
@@ -9,6 +9,7 @@
     "dist/types",
     "dist/src/**/*.dto.{js,d.ts,ts}",
     "dist/src/**/constants.{js,d.ts,ts}",
+    "dist/src/**/schema.{js,d.ts,ts}",
     "dist/src/**/exports.{js,d.ts,ts}",
     "dist/src/shared/dto-modification",
     "dist/src/shared/errors/dotType.ts",
@@ -110,10 +111,6 @@
       "require": "./dist/types/index.js",
       "default": "./dist/types/index.js",
       "types": "./dist/types/index.d.ts"
-    },
-    "./dto": {
-      "default": "./dist/src/prisma/dtos.dto.js",
-      "types": "./dist/src/prisma/dtos.dto.d.ts"
     }
   },
   "lint-staged": {
