+++ src/profile/dto/find-all-profile.dto.ts
@@ -1,27 +1,5 @@
-import { profileSchema } from '@/prisma/dtos.dto';
+import { findAllProfileResponseSchema } from '@/profile/schema/find-all-profile.schema';
 import { createZodDtoWithoutDate } from '@/shared/dto-modification/create-zod-dto-without-date';
-import { tagGroupSchema } from '@/tag-group/dto/tag-group.dto';
-import { tagSchema } from '@/tag/dto/tag.dto';
-import z from 'zod';
-
-export const findAllProfileResponseSchema = z.object({
-  profiles: z.array(
-    profileSchema.merge(
-      z.object({
-        tags: z.array(
-          tagSchema.merge(
-            z.object({
-              group: tagGroupSchema.pick({
-                color: true,
-                isExclusive: true,
-              }),
-            }),
-          ),
-        ),
-      }),
-    ),
-  ),
-});
 
 export class FindAllProfileResponseDto extends createZodDtoWithoutDate(
   findAllProfileResponseSchema,
