+++ src/profile/dto/find-by-id-profile.dto.ts
@@ -1,27 +1,5 @@
-import { locationSchema } from '@/location/dto/location.dto';
-import { profileSchema } from '@/prisma/dtos.dto';
+import { findByIdProfileResponseSchema } from '@/profile/schema/find-by-id-profile.schema';
 import { createZodDtoWithoutDate } from '@/shared/dto-modification/create-zod-dto-without-date';
-import { tagGroupSchema } from '@/tag-group/dto/tag-group.dto';
-import { tagSchema } from '@/tag/dto/tag.dto';
-import z from 'zod';
-
-export const findByIdProfileResponseSchema = profileSchema.merge(
-  z.object({
-    residenceLocation: locationSchema.nullable(),
-    birthLocation: locationSchema.nullable(),
-    tags: z.array(
-      tagSchema.merge(
-        z.object({
-          group: tagGroupSchema.pick({
-            id: true,
-            color: true,
-            isExclusive: true,
-          }),
-        }),
-      ),
-    ),
-  }),
-);
 
 export class FindByIdProfileResponseDto extends createZodDtoWithoutDate(
   findByIdProfileResponseSchema,
