+++ src/profile/dto/find-by-phone-number.dto.ts
@@ -1,7 +1,5 @@
-import { profileSchema } from '@/prisma/dtos.dto';
+import { findByPhoneNumberResponseSchema } from '@/profile/schema/find-by-phone-number.schema';
 import { createZodDtoWithoutDate } from '@/shared/dto-modification/create-zod-dto-without-date';
-
-export const findByPhoneNumberResponseSchema = profileSchema;
 
 export class FindByPhoneNumberResponseDto extends createZodDtoWithoutDate(
   findByPhoneNumberResponseSchema,
