+++ src/profile/exports.ts
@@ -1,12 +1,12 @@
-export * from './dto/create-profile.dto';
-export * from './dto/delete-profile.dto';
-export * from './dto/find-all-profile.dto';
-export * from './dto/find-by-date-range-profile.dto';
-export * from './dto/find-by-id-profile.dto';
-export * from './dto/find-by-phone-number.dto';
-export * from './dto/find-by-tag-groups-profile.dto';
-export * from './dto/find-by-tags-profile.dto';
-export * from './dto/find-trash.dto';
-export * from './dto/find-with-active-chat.dto';
-export * from './dto/profile.dto';
-export * from './dto/update-profile.dto';
+export * from './schema/create-profile.schema';
+export * from './schema/delete-profile.schema';
+export * from './schema/find-all-profile.schema';
+export * from './schema/find-by-date-range-profile.schema';
+export * from './schema/find-by-id-profile.schema';
+export * from './schema/find-by-phone-number.schema';
+export * from './schema/find-by-tag-groups-profile.schema';
+export * from './schema/find-by-tags-profile.schema';
+export * from './schema/find-trash.schema';
+export * from './schema/find-with-active-chat.schema';
+export * from './schema/profile.schema';
+export * from './schema/update-profile.schema';
