+++ src/profile/schema/delete-profile.schema.ts
@@ -1,0 +1,3 @@
+import { profileSchema } from '@/profile/schema/profile.schema';
+
+export const deleteProfileResponseSchema = profileSchema;
