+++ src/profile/schema/find-by-id-profile.schema.ts
@@ -1,0 +1,23 @@
+import { locationSchema } from '@/location/dto/location.dto';
+import { profileSchema } from '@/profile/schema/profile.schema';
+import { tagGroupSchema } from '@/tag-group/dto/tag-group.dto';
+import { tagSchema } from '@/tag/dto/tag.dto';
+import z from 'zod';
+
+export const findByIdProfileResponseSchema = profileSchema.merge(
+  z.object({
+    residenceLocation: locationSchema.nullable(),
+    birthLocation: locationSchema.nullable(),
+    tags: z.array(
+      tagSchema.merge(
+        z.object({
+          group: tagGroupSchema.pick({
+            id: true,
+            color: true,
+            isExclusive: true,
+          }),
+        }),
+      ),
+    ),
+  }),
+);
