+++ src/profile/schema/find-by-phone-number.schema.ts
@@ -1,0 +1,3 @@
+import { profileSchema } from '@/profile/schema/profile.schema';
+
+export const findByPhoneNumberResponseSchema = profileSchema;
