+++ src/profile/schema/find-by-tag-groups-profile.schema.ts
@@ -1,0 +1,13 @@
+import { profileSchema } from '@/profile/schema/profile.schema';
+import { tagSchema } from '@/tag/dto/tag.dto';
+import z from 'zod';
+
+export const findByTagGroupsProfileResponseSchema = z.object({
+  profiles: z.array(
+    profileSchema.merge(
+      z.object({
+        tags: z.array(tagSchema),
+      }),
+    ),
+  ),
+});
